+++ lib/view_models/date_time_cubit/date_time_state.dart
@@ -1,27 +1,26 @@
 part of 'date_time_cubit.dart';
 
-
-class DateTimeState extends Equatable {
-  final DateTime? dateTime;
-  final String? fieldError;
-
-  const DateTimeState({
-    this.dateTime,
-    this.fieldError,
-  });
-
-  bool get isValid => fieldError == null;
-
-  DateTimeState copyWith({
-    DateTime? dateTime,
-    String? fieldError,
-  }) {
-    return DateTimeState(
-      dateTime: dateTime ?? this.dateTime,
-      fieldError: fieldError,
-    );
-  }
-
-  @override
-  List<Object?> get props => [dateTime, fieldError];
-}
+// class DateTimeState extends Equatable {
+//   final DateTime? dateTime;
+//   final String? fieldError;
+//
+//   const DateTimeState({
+//     this.dateTime,
+//     this.fieldError,
+//   });
+//
+//   bool get isValid => fieldError == null;
+//
+//   DateTimeState copyWith({
+//     DateTime? dateTime,
+//     String? fieldError,
+//   }) {
+//     return DateTimeState(
+//       dateTime: dateTime ?? this.dateTime,
+//       fieldError: fieldError,
+//     );
+//   }
+//
+//   @override
+//   List<Object?> get props => [dateTime, fieldError];
+// }